Extract UntrustedDeviceDialog and fix e2ee icon
